moved website_fixes to a dedicated app
- I should not use models of base app
- website-fixes started to seem like a totally separate app, so had to move
- moved the templates, forms, views,
- models are left. IT WOULD WORK now if we renamed the database tables to the new app name instead of base_app, but in the next commit I will try to solve the model migration from one app to another in a proper way
