+++ lifeapi_apps/base_app/admin.py
@@ -1,7 +1,1 @@
 from django.contrib import admin
-from .models import WebsiteFix, WebsiteFixTag
-
-# Register your models here.
-
-admin.site.register(WebsiteFix)
-admin.site.register(WebsiteFixTag)
