+++ lifeapi_apps/base_app/templates/home.html
@@ -14,7 +14,4 @@
     <script type="text/javascript" src="https://day.lt/dienos_info?format=Dienos patarlė: [patarle][n]&charset=utf"></script>
     <br>
     <script type="text/javascript" src="https://day.lt/dienos_info?format=Dienos šventė: [svente][n]&charset=utf"></script>
-
-    
-    
-    {% endblock %}
+{% endblock %}
