+++ lifeapi_apps/website_fixes_app/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import WebsiteFix, WebsiteFixTag
+
+admin.site.register(WebsiteFix)
+admin.site.register(WebsiteFixTag)
