+++ lifeapi_apps/website_fixes_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class WebsiteFixesAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'lifeapi_apps.website_fixes_app'
