+++ lifeapi_apps/website_fixes_app/templates/website_fixes/home.html
@@ -1,0 +1,5 @@
+{% extends 'base.html' %}
+
+{% block content %}
+    <h1>Welcome to My Website Fixes</h1>
+{% endblock %}
