+++ lifeapi_project/urls.py
@@ -25,4 +25,5 @@ urlpatterns = [
     path("__debug__/", include("debug_toolbar.urls")),
     path("members/", include('django.contrib.auth.urls')),
     path("members/", include("lifeapi_apps.members_app.urls")),
+    path("website_fixes/", include("lifeapi_apps.website_fixes_app.urls")),
 ]
